LPD-32757 Move code to correct class
